SUBMARINE-658. Rename publick-api.ts to public-api.ts
